+++ src/store/index.ts
@@ -1,6 +1,0 @@
-export { imageModule } from '@/store/modules/image';
-export { loadingModule } from '@/store/modules/loading';
-export { piStatusModule } from '@/store/modules/piStatus';
-export { snackbarModule } from '@/store/modules/snackbar';
-export { userModule } from '@/store/modules/user';
-export { websocketModule } from '@/store/modules/websocket';