+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# <a href='https://github.com/mrjackwills/leafcast_vue/releases/tag/v0.9.5'>v0.9.5</a>
+### 2023-05-11
+
+### Chores
++ dependencies updated, [c34fcc33](https://github.com/mrjackwills/leafcast_vue/commit/c34fcc33a577e204f0a572a103ba1f66ca5ef3c6), [43c30e71](https://github.com/mrjackwills/leafcast_vue/commit/43c30e71fd1a98b7238e842004ae700c3b5bf43e), [608841cd](https://github.com/mrjackwills/leafcast_vue/commit/608841cd7334642253ee215669a6c7e63ec1d67f)
+
+### Features
++ typed defineProps, [de7bb355](https://github.com/mrjackwills/leafcast_vue/commit/de7bb3557b5eb726ec0b24c8c3686d856233a4c6)
+
+### Fixes
++ autoimport pinia store, [2035dcb4](https://github.com/mrjackwills/leafcast_vue/commit/2035dcb4f9289df35963a498e5c72b13145f79a6)
+
 # <a href='https://github.com/mrjackwills/leafcast_vue/releases/tag/v0.9.4'>v0.9.4</a>
 ### 2023-02-08